cloud_functions/fetch_openmeteo_weather/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.30.0
google-cloud-logging==3.11.0
requests>=2.28.0
